+++ Assets/Scripts/Encounters/Shop/ShopManager.cs
@@ -135,7 +135,7 @@ public class ShopManager : GenericSingleton<ShopManager>, IEncounterBuilder
             this.companionInShop = companionInShop;
             newCompanion = new Companion(companionInShop.companionType);
             if (ProgressManager.Instance.IsFeatureEnabled(AscensionType.DAMAGED_COMPANIONS)) {
-                newCompanion.combatStats.currentHealth -= ProgressManager.Instance.GetAscensionSO(AscensionType.DAMAGED_COMPANIONS).modificationValue;
+                newCompanion.combatStats.currentHealth -= (int)ProgressManager.Instance.GetAscensionSO(AscensionType.DAMAGED_COMPANIONS).modificationValue;
             }
             // companionToAdd is the final companion to add to your team :)
             Companion companionToAdd = newCompanion;
